# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a RBAC server.",
        "title": "RBAC GO API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/rbac/role": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC role"
                ],
                "summary": "Role 생성",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dblayer.RoleData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Role"
                        }
                    }
                }
            }
        },
        "/rbac/role/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC role"
                ],
                "summary": "Role 목록 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dblayer.RolesPage"
                        }
                    }
                }
            }
        },
        "/rbac/role/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC role"
                ],
                "summary": "Role 삭제",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제된 Role 데이터",
                        "schema": {
                            "$ref": "#/definitions/models.Role"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC role"
                ],
                "summary": "Role Update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update에 사용할 Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dblayer.RoleData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "수정된 Role 데이터",
                        "schema": {
                            "$ref": "#/definitions/models.Role"
                        }
                    }
                }
            }
        },
        "/rbac/role/{id}/permission": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC role"
                ],
                "summary": "Role에 속하는 Permission들의 id 목록 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dblayer.PermissionsOfRolePage"
                        }
                    }
                }
            }
        },
        "/rbac/role/{id}/subject": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC role"
                ],
                "summary": "Role에 속하는 Subject들의 id 목록 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dblayer.SubjectsOfRolePage"
                        }
                    }
                }
            }
        },
        "/rbac/subject-assignment": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC subjectAssignment"
                ],
                "summary": "SubjectAssignment 생성",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dblayer.SubjectAssignmentData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SubjectAssignment"
                        }
                    }
                }
            }
        },
        "/rbac/subject-assignment/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC subjectAssignment"
                ],
                "summary": "SubjectAssignment 삭제",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SubjectAssignment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제된 SubjectAssignment 데이터",
                        "schema": {
                            "$ref": "#/definitions/models.SubjectAssignment"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC subjectAssignment"
                ],
                "summary": "SubjectAssignment Update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SubjectAssignment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update에 사용할 Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dblayer.SubjectAssignmentData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "수정된 SubjectAssignment 데이터",
                        "schema": {
                            "$ref": "#/definitions/models.SubjectAssignment"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dblayer.PermissionOfRole": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "permission_id": {
                    "type": "integer"
                },
                "role_id": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                }
            }
        },
        "dblayer.PermissionsOfRolePage": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "previous": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dblayer.PermissionOfRole"
                    }
                }
            }
        },
...
        "dblayer.RoleData": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dblayer.RolesPage": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "previous": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Role"
                    }
                }
            }
        },
        "dblayer.SubjectAssignmentData": {
            "type": "object",
            "properties": {
                "role_id": {
                    "type": "integer"
                },
                "subject_id": {
                    "type": "integer"
                }
            }
        },
        "dblayer.SubjectsOfRolePage": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "previous": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SubjectAssignment"
                    }
                }
            }
        },
...
        "models.Role": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.SubjectAssignment": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "role_id": {
                    "type": "integer"
                },
                "subject_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}